[ci skip] moveonly: move ismine stuff to be a module of LegacyScriptPubKeyMan
